+++ pykeepass/kdbx_parsing/kdbx.py
@@ -1,14 +1,19 @@
-from construct import Struct, Switch, Bytes, Int16ul, RawCopy, this
+from construct import Struct, Switch, Bytes, Int16ul, RawCopy, Check, this
 from .kdbx3 import DynamicHeader as DynamicHeader3
 from .kdbx3 import Body as Body3
 from .kdbx4 import DynamicHeader as DynamicHeader4
 from .kdbx4 import Body as Body4
 
+# verify file signature
+def check_signature(ctx):
+    return ctx.sig1 == b'\x03\xd9\xa2\x9a' and ctx.sig2 == b'\x67\xFB\x4B\xB5'
+
 KDBX = Struct(
     "header" / RawCopy(
         Struct(
-            "magic1" / Bytes(4),
-            "magic2" / Bytes(4),
+            "sig1" / Bytes(4),
+            "sig2" / Bytes(4),
+            "sig_check" / Check(check_signature),
             "minor_version" / Int16ul,
             "major_version" / Int16ul,
             "dynamic_header" / Switch(